requirements.txt
ccxt
requests
python-dotenv
numpy
pandas
matplotlib
scikit-learn
TA-Lib
torch
aiohttp
cachetools
deepseek-ai
